+++ billy/site/browse/views.py
@@ -27,32 +27,32 @@ def browse_index(request, template='billy/index.html'):
         row['id'] = meta['_id']
         row['name'] = meta['name']
 
-        counts = db.counts.find_one({'_id': row['id']})
-        if counts:
-            counts = counts['value']
-            row['bills'] = counts['bills']
-            row['votes'] = counts['votes']
-            row['versions'] = counts['versions']
-            if counts['actions']:
-                row['typed_actions'] = (float(counts['categorized']) /
-                                        counts['actions_this_term'] * 100)
-            if counts['bills']:
-                row['subjects'] = (float(counts['subjects']) /
-                                   counts['bills'] * 100)
-            if counts['sponsors']:
-                row['sponsor_ids'] = (float(counts['idd_sponsors']) /
-                                      counts['sponsors'] * 100)
-            if counts['voters']:
-                row['voter_ids'] = (float(counts['idd_voters']) /
-                                    counts['voters'] * 100)
+        bill_stats = db.bill_stats.find_one({'_id': row['id']})
+        if bill_stats:
+            bill_stats = bill_stats['value']
+            row['bills'] = bill_stats['bills']
+            row['votes'] = bill_stats['votes']
+            row['versions'] = bill_stats['versions']
+            if bill_stats['actions']:
+                row['typed_actions'] = (float(bill_stats['categorized']) /
+                                        bill_stats['actions_this_term'] * 100)
+            if bill_stats['bills']:
+                row['subjects'] = (float(bill_stats['subjects']) /
+                                   bill_stats['bills'] * 100)
+            if bill_stats['sponsors']:
+                row['sponsor_ids'] = (float(bill_stats['idd_sponsors']) /
+                                      bill_stats['sponsors'] * 100)
+            if bill_stats['voters']:
+                row['voter_ids'] = (float(bill_stats['idd_voters']) /
+                                    bill_stats['voters'] * 100)
 
-        com_counts = db.committee_counts.find_one({'_id': row['id']})
-        if com_counts:
-            com_counts = com_counts['value']
-            row['committees'] = com_counts['committees']
-            if com_counts['members']:
-                row['member_ids'] = (float(com_counts['idd_members']) /
-                                     com_counts['members'] * 100)
+        com_stats = db.committee_stats.find_one({'_id': row['id']})
+        if com_stats:
+            com_stats = com_stats['value']
+            row['committees'] = com_stats['committees']
+            if com_stats['members']:
+                row['member_ids'] = (float(com_stats['idd_members']) /
+                                     com_stats['members'] * 100)
 
         if row['id'] != 'total':
             level = meta['level']